update setup scripts, add oh-my-zsh

+++ conf/install.sh
@@ -2,21 +2,29 @@
 
 set -e
 
-cp ~/.vimrc ./vimrc.backup 2> /dev/null
+if [[ -e ~/.vimrc ]]; then
+	cp ~/.vimrc ./vimrc.backup
+fi
+if [[ -e ~/.zshrc ]]; then
+	cp ~/.zshrc ./zshrc.backup
+fi
+if [[ -e ~/.zsh_plugins.txt ]]; then
+	cp ~/.zsh_plugins.txt ./zsh_plugins.txt.backup 2> /dev/null
+fi
+if [[ -e ~/.gdbinit ]]; then
+	cp ~/.gdbinit ./gdbinit.backup 2> /dev/null
+fi
+if [[ -e ~/.lldbinit ]]; then
+	cp ~/.lldbinit ./lldbinit.backup 2> /dev/null
+fi
+if [[ -e ~/.pdbrc ]]; then
+	cp ~/.pdbrc ./pdbrc.backup 2> /dev/null
+fi
+
 cp vimrc ~/.vimrc
-
-cp ~/.zshrc ./zshrc.backup 2> /dev/null
 cp zshrc ~/.zshrc
-
-cp ~/.zsh_plugins.txt ./zsh_plugins.txt.backup 2> /dev/null
 cp zsh_plugins.txt ~/.zsh_plugins.txt
-
-cp ~/.gdbinit ./gdbinit.backup 2> /dev/null
 cp gdbinit ~/.gdbinit
-
-cp ~/.lldbinit ./lldbinit.backup 2> /dev/null
 cp lldbinit ~/.lldbinit
-
-cp ~/.pdbrc ./pdbrc.backup 2> /dev/null
 cp pdbrc ~/.pdbrc
